refactor: changes due to code review  [WTEL-4046]
